extract ndex network property editor to it's own component
