+++ src/components/SummaryPanel/NetworkPropertyPanel.tsx
@@ -3,13 +3,9 @@ import {
   Tooltip,
   IconButton,
   Box,
-  Chip,
   Theme,
   Typography,
   Divider,
-  Paper,
-  Popover,
-  TextField,
   Dialog,
   DialogContent,
   DialogContentText,
@@ -18,27 +14,14 @@ import {
   DialogActions,
   ButtonGroup,
 } from '@mui/material'
-import { debounce } from 'lodash'
 import { blueGrey } from '@mui/material/colors'
 import { useTheme } from '@mui/material/styles'
 import EditIcon from '@mui/icons-material/Edit'
 import CircleIcon from '@mui/icons-material/Circle'
 
-import { RichTextEditor, Link } from '@mantine/tiptap'
-import { useEditor } from '@tiptap/react'
-import Highlight from '@tiptap/extension-highlight'
-import StarterKit from '@tiptap/starter-kit'
-import Underline from '@tiptap/extension-underline'
-import TextAlign from '@tiptap/extension-text-align'
-import Superscript from '@tiptap/extension-superscript'
-import SubScript from '@tiptap/extension-subscript'
-
 import { IdType } from '../../models/IdType'
 import { NdexNetworkSummary } from '../../models/NetworkSummaryModel'
 import { useWorkspaceStore } from '../../store/WorkspaceStore'
-import { useNetworkSummaryStore } from '../../store/NetworkSummaryStore'
-import NdexNetworkPropertyTable from './NdexNetworkPropertyTable'
-import { removePTags } from '../../utils/remove-p-tags'
 import { useViewModelStore } from '../../store/ViewModelStore'
 import { useHcxValidatorStore } from '../../features/HierarchyViewer/store/HcxValidatorStore'
 import { PublishedWithChanges, WarningAmberOutlined } from '@mui/icons-material'
@@ -47,6 +30,7 @@ import { validateHcx } from '../../features/HierarchyViewer/model/impl/hcxValida
 import { useTableStore } from '../../store/TableStore'
 import { useMessageStore } from '../../store/MessageStore'
 import CheckCircleOutlineIcon from '@mui/icons-material/CheckCircleOutline'
+import { NetworkPropertyEditor } from './NdexNetworkPropertyEditor'
 
 interface NetworkPropertyPanelProps {
   summary: NdexNetworkSummary
@@ -122,8 +106,6 @@ export const NetworkPropertyPanel = ({
     (state) => state.setCurrentNetworkId,
   )
 
-  const updateNetworkSummary = useNetworkSummaryStore((state) => state.update)
-
   const networkModified =
     useWorkspaceStore((state) => state.workspace.networkModified[id]) ?? false
 
@@ -140,24 +122,6 @@ export const NetworkPropertyPanel = ({
     (state) => state.validationResults,
   )
   const validationResult = validationResults?.[id]
-
-  const editor = useEditor({
-    onUpdate: debounce(({ editor }) => {
-      updateNetworkSummary(summary.externalId, {
-        description: editor.getHTML(),
-      })
-    }, 200),
-    extensions: [
-      StarterKit,
-      Underline,
-      Link,
-      Superscript,
-      SubScript,
-      Highlight,
-      TextAlign.configure({ types: ['heading', 'paragraph'] }),
-    ],
-    content: removePTags(summary.description ?? ''),
-  })
 
   return (
     <>
@@ -264,106 +228,12 @@ export const NetworkPropertyPanel = ({
             </Button>
           </DialogActions>
         </Dialog>
-        <Popover
-          open={editNetworkSummaryAnchorEl !== undefined}
+
+        <NetworkPropertyEditor
           anchorEl={editNetworkSummaryAnchorEl}
+          summary={summary}
           onClose={hideEditNetworkSummaryForm}
-          anchorOrigin={{
-            vertical: 'top',
-            horizontal: 'right',
-          }}
-        >
-          <Paper
-            sx={{
-              p: 2,
-              width: 800,
-              height: 800,
-              overflowY: 'scroll',
-            }}
-          >
-            <Chip
-              sx={{ p: 1, mb: 2 }}
-              size="small"
-              label={
-                <Typography variant="caption">{summary.visibility}</Typography>
-              }
-            />
-            <Box sx={{ display: 'flex', justifyContent: 'space-between' }}>
-              <TextField
-                size="small"
-                label="Name"
-                sx={{ width: '60%', mr: 1, fontSize: 12 }}
-                value={summary.name}
-                onChange={(e) => {
-                  updateNetworkSummary(summary.externalId, {
-                    name: e.target.value,
-                  })
-                }}
-              ></TextField>
-              <TextField
-                size="small"
-                label="Version"
-                sx={{ width: '20%', fontSize: 12 }}
-                value={summary.version}
-                onChange={(e) =>
-                  updateNetworkSummary(summary.externalId, {
-                    version: e.target.value,
-                  })
-                }
-              />
-            </Box>
-
-            <Typography sx={{ ml: 1.5, pt: 1 }} gutterBottom>
-              Description
-            </Typography>
-            <RichTextEditor editor={editor}>
-              <RichTextEditor.Toolbar sticky stickyOffset={60}>
-                <RichTextEditor.ControlsGroup>
-                  <RichTextEditor.Bold />
-                  <RichTextEditor.Italic />
-                  <RichTextEditor.Underline />
-                  <RichTextEditor.Strikethrough />
-                  <RichTextEditor.ClearFormatting />
-                  <RichTextEditor.Highlight />
-                  <RichTextEditor.Code />
-                </RichTextEditor.ControlsGroup>
-
-                <RichTextEditor.ControlsGroup>
-                  <RichTextEditor.H1 />
-                  <RichTextEditor.H2 />
-                  <RichTextEditor.H3 />
-                  <RichTextEditor.H4 />
-                </RichTextEditor.ControlsGroup>
-
-                <RichTextEditor.ControlsGroup>
-                  <RichTextEditor.Blockquote />
-                  <RichTextEditor.Hr />
-                  <RichTextEditor.BulletList />
-                  <RichTextEditor.OrderedList />
-                  <RichTextEditor.Subscript />
-                  <RichTextEditor.Superscript />
-                </RichTextEditor.ControlsGroup>
-
-                <RichTextEditor.ControlsGroup>
-                  <RichTextEditor.Link />
-                  <RichTextEditor.Unlink />
-                </RichTextEditor.ControlsGroup>
-
-                <RichTextEditor.ControlsGroup>
-                  <RichTextEditor.AlignLeft />
-                  <RichTextEditor.AlignCenter />
-                  <RichTextEditor.AlignJustify />
-                  <RichTextEditor.AlignRight />
-                </RichTextEditor.ControlsGroup>
-              </RichTextEditor.Toolbar>
-
-              <RichTextEditor.Content />
-            </RichTextEditor>
-
-            <Divider sx={{ mt: 2, mb: 1 }} />
-            <NdexNetworkPropertyTable />
-          </Paper>
-        </Popover>
+        />
       </Box>
     </>
   )
